requirements.txt
# ============================================
# DeepTutor Python Dependencies
# ============================================
# Python >= 3.10 required
# Install: pip install -r requirements.txt

# ============================================
# Core dependencies
# ============================================
python-dotenv>=1.0.0
PyYAML>=6.0
tiktoken>=0.5.0
jinja2>=3.1.0              # Template engine for search consolidation
minja>=0.1.0               # Lightweight Jinja renderer for chat templates

# ============================================
# HTTP and API clients
# ============================================
openai>=1.30.0
anthropic>=0.30.0          # Anthropic Claude API
perplexityai>=0.1.0
dashscope>=1.14.0
redis>=5.0.0
httpx>=0.27.0
urllib3>=2.2.1

# ============================================
# Async support
# ============================================
nest_asyncio>=1.5.8
tenacity>=8.0.0
aiofiles>=23.0.0

# ============================================
# Web framework and server
# ============================================
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
pydantic>=2.0.0

# ============================================
# RAG and knowledge base
# ============================================
raganything>=0.1.0
lightrag>=0.1.0b6
docling>=2.31.0                # Document parser for Office/HTML (alternative to MinerU)

# ============================================
# LlamaIndex
# ============================================
llama-index>=0.14.12

# ============================================
# Academic and research tools
# ============================================
arxiv>=2.0.0

# ============================================
# Scientific computing (for RAG and code execution)
# ============================================
numpy>=1.24.0,<2.0.0  # NumPy 1.24+ required for array API, <2.0 for compatibility
# matplotlib>=3.7.0  # Uncomment if needed for plotting
# scipy>=1.11.0      # Uncomment if needed for scientific computing
# pandas>=2.0.0      # Uncomment if needed for data analysis

# PDF parsing
PyMuPDF>=1.26.0      # For PDF text extraction (fast and reliable)

# Vector search (optional but recommended for llamaindex pipeline)
# faiss-cpu>=1.7.4   # Uncomment for CPU-only FAISS support (faster search)
# faiss-gpu>=1.7.4   # Uncomment if you have GPU and want GPU acceleration

# ============================================
# Development and testing tools
# ============================================
pytest>=7.0.0              # Testing framework
pytest-asyncio>=0.22.0     # Async test support for pytest
hypothesis>=6.90.0
pre-commit>=3.0.0
safety<3.0.0               # Pin to stable 2.x version, 3.0.0 is broken
bandit>=1.8.0
